requirements.txt
acres==0.3.0
annotated-types==0.7.0
asgiref==3.8.1
astroid==3.3.10
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.11
dill==0.4.0
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
etelemetry==0.3.1
filelock==3.18.0
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
isort==6.0.1
itypes==1.2.0
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
looseversion==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
murmurhash==1.0.12
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pillow==11.2.1
platformdirs==4.3.8
preshed==3.0.9
prov==2.0.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
puremagic==1.29
pydantic==2.11.4
pydantic_core==2.33.2
pydot==4.0.0
Pygments==2.19.1
PyJWT==2.9.0
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
PyMuPDF==1.25.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
requests==2.32.3
rich==14.0.0
scipy==1.15.3
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
thinc==8.3.6
tomlkit==0.13.2
tqdm==4.67.1
traits==7.0.2
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.